+++ Bet apps/src/chorke/proprietary/bet/apps/core/bets/BetOverUnder.java
@@ -4,17 +4,19 @@
  */
 package chorke.proprietary.bet.apps.core.bets;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author Chorke
  */
 public class BetOverUnder extends Bet{
-    public final double total;
-    public final double over;
-    public final double under;
+    public final BigDecimal total;
+    public final BigDecimal over;
+    public final BigDecimal under;
     public final String description;
 
-    public BetOverUnder(String betCompany, double total, double over, double under, String description) {
+    public BetOverUnder(String betCompany, BigDecimal total, BigDecimal over, BigDecimal under, String description) {
         super(betCompany);
         this.total = total;
         this.over = over;
