+++ Bet apps/src/chorke/proprietary/bet/apps/core/calculators/CumulativeYieldProperties.java
@@ -4,6 +4,7 @@
  */
 package chorke.proprietary.bet.apps.core.calculators;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -17,11 +18,11 @@ public class CumulativeYieldProperties extends YieldProperties{
         super();
     }
     
-    public CumulativeYieldProperties(List<Double> scale, String betCompany) {
+    public CumulativeYieldProperties(List<BigDecimal> scale, String betCompany) {
         super(scale, betCompany);
     }
 
-    public CumulativeYieldProperties(List<Double> scale, String betCompany, Map<String, Object> propeerties) {
+    public CumulativeYieldProperties(List<BigDecimal> scale, String betCompany, Map<String, Object> propeerties) {
         super(scale, betCompany, propeerties);
     }
